requirements.txt
bcrypt==3.1.6
blinker==1.4
cffi==1.12.3
Click==7.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.4.0
flask-whooshalchemy3==0.1.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pbr==3.1.1
pycparser==2.19
six==1.12.0
SQLAlchemy==1.1.13
Werkzeug==0.15.4
Whoosh==2.7.4
WTForms==2.2.1
